requirements.txt
openai==1.97.1
streamlit==1.47.1
python-dotenv==1.1.1
requests==2.32.4
